requirements/stack.txt
pandas
awswrangler
aws-cdk.core
aws-cdk.aws_iam
aws-cdk.aws_sqs
aws-cdk.aws_ses
aws-cdk.aws_ses_actions
aws-cdk.aws_sns
aws-cdk.aws_sns_subscriptions
aws-cdk.aws_s3
aws_cdk.aws_dynamodb
aws_cdk.aws_lambda_python
aws_cdk.aws_lambda
aws_cdk.aws_s3_notifications
aws_cdk.aws_lambda_python
aws_cdk.aws_glue
aws_cdk.aws_s3_deployment
aws-cdk.aws-lakeformation
aws_cdk.aws_lambda_python
aws_cdk.aws_ecr_assets
aws_cdk.assertions
